refactor(solution-finder): delete TCO model; waste + TRAILS pricing survive
